rename transfer status interfaces, cleanup imports
