+++ packages/web/integrations/bridges/axelar/tokens.ts
@@ -1,5 +1,6 @@
 import { EthereumChainInfo } from "~/integrations/bridge-info";
-import { SourceChainTokenConfig } from "~/integrations/bridges/axelar/types";
+
+import { SourceChainTokenConfig } from "./types";
 
 const IS_TESTNET = process.env.NEXT_PUBLIC_IS_TESTNET === "true";
