+++ packages/web/stores/root.ts
@@ -36,9 +36,11 @@ import {
 } from "~/config";
 import { AssetLists } from "~/config/generated/asset-lists";
 import { ChainList } from "~/config/generated/chain-list";
-import { AxelarTransferStatusSource } from "~/integrations/bridges/axelar/axelar-transfer-status-source";
-import { SkipTransferStatusSource } from "~/integrations/bridges/skip/skip-transfer-status-source";
-import { SquidTransferStatusSource } from "~/integrations/bridges/squid/squid-transfer-status-source";
+import {
+  AxelarTransferStatusProvider,
+  SkipTransferStatusProvider,
+  SquidTransferStatusProvider,
+} from "~/integrations/bridges";
 import { ObservableAssets } from "~/stores/assets";
 import { makeIndexedKVStore, makeLocalStorageKVStore } from "~/stores/kv-store";
 import { NavBarStore } from "~/stores/nav-bar";
@@ -247,9 +249,9 @@ export class RootStore {
       this.chainStore.osmosis.chainId,
       makeLocalStorageKVStore("nonibc_transfer_history"),
       [
-        new AxelarTransferStatusSource(IS_TESTNET ? "testnet" : "mainnet"),
-        new SquidTransferStatusSource(IS_TESTNET ? "testnet" : "mainnet"),
-        new SkipTransferStatusSource(IS_TESTNET ? "testnet" : "mainnet"),
+        new AxelarTransferStatusProvider(IS_TESTNET ? "testnet" : "mainnet"),
+        new SquidTransferStatusProvider(IS_TESTNET ? "testnet" : "mainnet"),
+        new SkipTransferStatusProvider(IS_TESTNET ? "testnet" : "mainnet"),
       ]
     );
